test: split extension tests by area
